MySimpleUtilities/Misc.cs
using System;
using System.IO;
using System.Security.Cryptography;
using System.Text;

namespace MySimpleUtilities
{
    public class Misc
    {
        public string ComputeMD5Hash(string _s)
        {
            var md5 = MD5.Create();
            byte[] hash = md5.ComputeHash(Encoding.UTF8.GetBytes(_s));
            StringBuilder sb = new StringBuilder();
            for (int i = 0; i < hash.Length; i++)
            {
                sb.Append(hash[i].ToString("X2"));
            }
            return sb.ToString();
        }
        
        public string ComputeSHA256Hash(string s)
        {
            var sha = SHA256.Create();
            byte[] hash = sha.ComputeHash(Encoding.UTF8.GetBytes(s));
            var sb = new StringBuilder();
            for (int i = 0; i < hash.Length; i++)
            {
                sb.Append(hash[i].ToString("X2"));
            }
            return sb.ToString();
        }

        public string GetSpecialFolderPath(System.Environment.SpecialFolder _specialFolderEnum)
        {
            return Environment.GetFolderPath(_specialFolderEnum);
        }

        public Stream GenerateStreamFromString(string s)
        {
            MemoryStream stream = new MemoryStream();
            StreamWriter writer = new StreamWriter(stream);
            writer.Write(s);
            writer.Flush();
            stream.Position = 0;
            return stream;
        }
    }
}
